force floored vals to int

+++ dipy/direction/probabilistic_direction_getter.pyx
@@ -120,9 +120,9 @@ cdef class ProbabilisticDirectionGetter(PmfGenDirectionGetter):
 
         ## find max cosine similarity from precomputed angle array
         ## point has to go from mm to ijk? - _map_to_voxel / _to_voxel_coordinates
-        p1 = np.floor(point[0])
-        p2 = np.floor(point[1])
-        p3 = np.floor(point[2])
+        p1 = np.floor(point[0]).astype("int") 
+        p2 = np.floor(point[1]).astype("int") 
+        p3 = np.floor(point[2]).astype("int") 
         coss = self.cos_mat[p1, p2, p3]
         print("ijk: " + str(p1) + ", " + str(p2) + ", " + str(p3) + "; coss: ", + str(coss))
         #coss = self.cos_mat[ 74, 87, 73 ]
